// types/country.ts
export interface CountryType {
  id?: number;
  name?: string;
  iso3?: string;
  iso2?: string;
  numeric_code?: string;
  phone_code?: string | number;
  capital?: string;
  currency?: string;
  currency_name?: string;
  currency_symbol?: string;
  tld?: string;
  native?: string;
  region?: string;
  subregion?: string;
  latitude?: string;
  longitude?: string;
  emoji?: string;
}